+++ chatimage-spigot/src/main/java/me/chickensaysbak/chatimage/spigot/PluginSpigot.java
@@ -140,12 +140,12 @@ public class PluginSpigot extends JavaPlugin implements Listener, PluginAdapter 
     }
 
     @Override
-    public void runAsyncTaskLater(Runnable task, int ticks) {
+    public void runAsyncTaskLater(Runnable task, long ticks) {
         getServer().getScheduler().runTaskLaterAsynchronously(this, task, ticks).getTaskId();
     }
 
     @Override
-    public void runTaskLater(Runnable task, int ticks) {
+    public void runTaskLater(Runnable task, long ticks) {
         getServer().getScheduler().runTaskLater(this, task, ticks).getTaskId();
     }
